\name{GGMridge-package}
\alias{GGMridge}
\docType{package}
\title{ Gaussian Graphical Models Using Ridge Penalty Followed by
        Thresholding and Reestimation
}
\description{ Estimation of partial correlation matrix using ridge 
              penalty followed by thresholding and reestimation. 
              Under multivariate Gaussian assumption, the matrix 
              constitutes an Gaussian graphical model (GGM).}
\details{
\tabular{ll}{
Package: \tab GGMridge\cr
Type: \tab Package\cr
Version: \tab 1.1\cr
Date: \tab 2016-02-01\cr
License: \tab  GPL-2 \cr
}
}
\author{Min Jin Ha}
\references{
Ha, M. J. and Sun, W. 
(2014).
Partial correlation matrix estimation using ridge penalty followed 
by thresholding and re-estimation.
Biometrics, 70, 762--770.
}
\keyword{package}
