% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ksStat.R
\name{ksStat}
\alias{ksStat}
\title{The Kolmogorov-Smirnov Statistic for p-Values}
\usage{
ksStat(p)
}
\arguments{
\item{p}{A numeric vector with p-values.}
}
\value{
Kolmogorov-Smirnov statistic
}
\description{
Calculates the Kolmogorov-Smirnov statistic for p-values
}
\examples{
 p <- stats::runif(100)
 ksStat(p = p)
 ks.test(p, y = "punif") # compare with ks.test
 
}
\author{
Min Jin Ha
}
